Add `merge?` option to `Telemetrex.span`
Initially we went the route of following `:telemetry.span` by not
merging the start and stop metadata. In practice we found that we often
wanted them to be merged and there was friction in doing it manually.
Turns out `:telemetry` made this decision to be more performant in high
throughput metrics where the merging would have a cost.

  - [ref](https://github.com/beam-telemetry/telemetry/pull/77)

By adding an option to turn merging on or off we can provide the best of
both worlds and still give users the ability to control where it is
needed or not.

RESOLVES: #19